move more video cache logic to videoCache.js
